Add ScopeList. Move Memory specific stuff in own Memory package in Dgame.Core. Clean up. Switch from manual GC deallocation in TileMap to ScopeList / destroy.

+++ Graphics/Font.d
@@ -8,7 +8,7 @@ private {
 	
 	import derelict.sdl2.ttf;
 	
-	import Dgame.Core.SmartPointer.Shared;
+	import Dgame.Core.Memory.SmartPointer.Shared;
 }
 
 ///version = Develop;
@@ -18,7 +18,7 @@ private Font*[] _finalizer;
 void _finalizeFont() {
 	debug writefln("Finalize Font (%d)", _finalizer.length);
 	
-	for (uint i = 0; i < _finalizer.length; i++) {
+	for (size_t i = 0; i < _finalizer.length; i++) {
 		if (_finalizer[i]) {
 			debug writefln(" -> Font finalized: %d", i);
